All controllers besides payment and discount checked; Dependency injection implemented; Input validation enforced; Endpoints' responses improved

+++ Controllers/OrdersController.cs
@@ -33,18 +33,51 @@ namespace RestaurantSystem.Controllers
         [Authorize(Roles = "RestaurantManager, Customer, RestaurantEveryDayUse")]
         public async Task<IActionResult> GetOrder([FromQuery] string restaurantId, [FromQuery] DateTime date, [FromQuery] string userId)
         {
+            IEnumerable<Order> orders = new List<Order>();
+
+            //Customer can not see another customers orders
             var currentUserEmail = HttpContext.User.Claims.FirstOrDefault(claim => claim.Type == System.Security.Claims.ClaimTypes.Email).Value;
             var userRole = HttpContext.User.Claims.FirstOrDefault(claim => claim.Type == System.Security.Claims.ClaimTypes.Role).Value;
-
-            if ((userRole == "Customer" && !await _permissionValidation.isUserTheSameAsync(userId, currentUserEmail)) ||
-                (userRole == "RestaurantManager"  && !await _permissionValidation.isManagerRestaurantOwnerAsync(restaurantId, currentUserEmail)) ||
-                ( userRole == "RestaurantEveryDayUse") && !await _permissionValidation.isEveryDayUseAccountRestaurantsOwnershipAsync(restaurantId, currentUserEmail))
+            if (userRole == "Customer")
             {
-                
-                BadRequest();
+                if (userId == null)
+                {
+                    return BadRequest(new { Error = "User id required" });
+                }
+                else
+                {
+                    if (!await _permissionValidation.isUserTheSameAsync(userId, currentUserEmail))
+                    {
+                        return Unauthorized(new { Error = "Can not retrieve another costomers orders" });
+                    }
+                    else
+                    {
+                        orders = await _orderRepository.GetAllAsync(null, DateTime.MinValue, userId);
+                    }
+                }
+            }
+            else if (userRole == "RestaurantManager" || userRole == "RestaurantEveryDayUse")
+            {
+                if (restaurantId == null || date == DateTime.MinValue)
+                {
+                    return BadRequest(new { Error = "RestaurantId and date required" });
+                }
+                else
+                {
+                    //Manager and RestaurantEveryDayUse can not retrieve another restaurants bookings
+                    if ((userRole == "RestaurantManager" && !await _permissionValidation.isManagerRestaurantOwnerAsync(restaurantId, currentUserEmail)) ||
+                        (userRole == "RestaurantEveryDayUse" && !await _permissionValidation.isEveryDayUseAccountRestaurantsOwnershipAsync(restaurantId, currentUserEmail)))
+                    {
+
+                        return Unauthorized(new { Error = "Can not retrieve orders of another restaurant" });
+                    }
+                    else
+                    {
+                        orders = await _orderRepository.GetAllAsync(restaurantId, date, null);
+                    }
+                }
             }
 
-            var orders = await _orderRepository.GetAllAsync(restaurantId, date, userId);
             return Ok(orders);
         }
 
@@ -53,24 +86,36 @@ namespace RestaurantSystem.Controllers
         [Authorize(Roles = "RestaurantManager, Customer, RestaurantEveryDayUse")]
         public async Task<ActionResult<Order>> GetOrder(string id)
         {
+            if (!await _orderRepository.IfExist(id))
+            {
+                return NotFound(new { Error = "Order not found" });
+            }
+
             var order = await _orderRepository.GetAsync(id);
 
             var currentUserEmail = HttpContext.User.Claims.FirstOrDefault(claim => claim.Type == System.Security.Claims.ClaimTypes.Email).Value;
             var userRole = HttpContext.User.Claims.FirstOrDefault(claim => claim.Type == System.Security.Claims.ClaimTypes.Role).Value;
-
-            if (userRole == "Customer" && currentUserEmail != order.Customer.Email)
+            if (userRole == "Customer")
             {
-                return BadRequest();
+                if (!await _permissionValidation.isUserTheSameAsync(order.Customer.SystemId, currentUserEmail))
+                {
+                    return Unauthorized(new { Error = "Can not retrieve another costomers booking" });
+                }
             }
-
-            if (order == null)
+            else if (userRole == "RestaurantManager" || userRole == "RestaurantEveryDayUse")
             {
-                return NotFound();
+                if ((userRole == "RestaurantManager" && !await _permissionValidation.isManagerRestaurantOwnerAsync(order.Restaurant.Id, currentUserEmail)) ||
+                    (userRole == "RestaurantEveryDayUse" && !await _permissionValidation.isEveryDayUseAccountRestaurantsOwnershipAsync(order.Restaurant.Id, currentUserEmail)))
+                {
+
+                    return Unauthorized(new { Error = "Can not retrieve orders of another restaurant" });
+                }
             }
 
             return Ok(order);
         }
 
+        /*
         // PUT: api/Orders/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
@@ -128,55 +173,70 @@ namespace RestaurantSystem.Controllers
 
             return NoContent();
         }
+        */
 
         // POST: api/Orders
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
-        [Authorize(Roles = "Customer")]
+        [Authorize(Roles = "RestaurantManager, Customer")]
         public async Task<ActionResult<Order>> PostOrder(OrderRequest orderRequest)
         {
             var currentUserEmail = HttpContext.User.Claims.FirstOrDefault(claim => claim.Type == System.Security.Claims.ClaimTypes.Email).Value;
             var userRole = HttpContext.User.Claims.FirstOrDefault(claim => claim.Type == System.Security.Claims.ClaimTypes.Role).Value;
 
-            if (userRole == "Customer")
+            if (!await _permissionValidation.isUserTheSameAsync(orderRequest.Customer, currentUserEmail))
             {
-                orderRequest.Customer = currentUserEmail;
+                return Unauthorized(new { Error = "Can not make an order for another user" });
             }
 
-            var order = await _orderRepository.ConvertAlterOrderRequest(orderRequest);
+            var order = await _orderRepository.ConvertAlterOrderRequest(orderRequest, null);
             if (order == null)
             {
-                return BadRequest();
+                return NotFound(new { Error = "One of order dependencies not found" });
             }
 
-            if (userRole == "RestaurantManager" || userRole == "RestaurantEveryDayUse")
+            foreach (var orderLine in orderRequest.OrderLines) 
             {
-                if (!await _permissionValidation.isManagerRestaurantOwnerAsync(orderRequest.Restaurant, currentUserEmail))
+                if(!await _permissionValidation.isDishRestaurantOwnershipAsync(orderLine.Dish, orderRequest.Restaurant))
                 {
-                    return Unauthorized();
+                    return BadRequest(new { Error = "One of chosen dishes does not belong to the chosen restaurant" });
                 }
-            }
-            var saved = await _orderRepository.CreateAsync(order);
+            };
 
-            if (saved == null)
+            if (!_orderRepository.ifOrderlinesUnique(orderRequest.OrderLines))
             {
-                return BadRequest();
+                return BadRequest(new { Error = "The same dish is in more that one orderline" });
             }
+
+            if (!_orderRepository.ifOrderlinesQuantityQuantityCorrect(orderRequest.OrderLines))
+            {
+                return BadRequest(new { Error = "Quantity should alway be between 1 and 99" });
+            }
+
+            var saved = await _orderRepository.CreateAsync(order);
 
             return CreatedAtAction("GetOrder", new { id = saved.Id }, saved);
         }
 
         // DELETE: api/Orders/5
         [HttpDelete("{id}")]
-        [Authorize(Roles = "RestaurantManager, Customer, RestaurantEveryDayUse")]
+        [Authorize(Roles = "RestaurantManager, Customer")]
         public async Task<IActionResult> DeleteOrder(string id)
         {
-            var returnedOrder = await _orderRepository.DeleteAsync(id);
-
-            if (returnedOrder == null)
+            var userRole = HttpContext.User.Claims.FirstOrDefault(claim => claim.Type == System.Security.Claims.ClaimTypes.Role).Value;
+            var currentUserEmail = HttpContext.User.Claims.FirstOrDefault(claim => claim.Type == System.Security.Claims.ClaimTypes.Email).Value;
+            if (!await _orderRepository.IfExist(id))
             {
-                return NotFound();
+                return NotFound(new { Error = "Order not found" });
             }
+
+            var order = await _orderRepository.GetAsync(id);
+            if (!await _permissionValidation.isUserTheSameAsync(order.Customer.SystemId, currentUserEmail))
+            {
+                return Unauthorized(new { Error = "Can not delete another user's order" });
+            }
+
+            var returnedOrder = await _orderRepository.DeleteAsync(id);
 
             return NoContent();
         }
